+++ resources/sql/queries.sql
@@ -33,3 +33,15 @@ WHERE id = :id
 -- :name get-users :? :*
 -- :doc get all users
 SELECT * FROM users
+
+-- :name user-by-name :? :1
+-- :doc retrieves a user record given the name
+SELECT * FROM users
+where name = :name
+
+-- :name update-tokens-by-name! :! :n
+-- :doc updates an existing user's tokens, key is name.
+UPDATE users
+SET access = :access_token, refresh = :refresh_token, userid = :userid,
+    updated_at = now()
+where name = :name